worker-app/requirements.txt
fastapi
uvicorn
pytesseract
opencv-python-headless
Pillow
numpy